requirements.txt
chatterbox-tts
torch
torchaudio
flask
waitress